vision-chat-py/pyproject.toml
[project]
name = "vision-chat-langchain-example"
version = "0.1.0"
description = "Python vision chat example for an OpenAI-compatible API using LangChain."
requires-python = ">=3.11"
dependencies = [
  "langchain-openai>=0.3.35",
  "python-dotenv>=1.0.0",
]

[tool.uv]
package = false
